+++ CHANGELOG.md
@@ -4,13 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://gitlab.devsep.com/septech-php/septech-laravel/snowflake/-/compare/v1.0.2...master)
 
-## [Unreleased](https://gitlab.devsep.com/septech-php/septech-laravel/snowflake/-/compare/v1.0.0...master)
+## [v1.0.2 - 2020-09-01](https://github.com/SepteniTechnology/laravel-snowflake/compare/v1.0.1...v1.0.2)
+### Changed
+- Update URL to github
 
-## [v1.0.1 - 2020-08-24](https://gitlab.devsep.com/septech-php/septech-laravel/snowflake/-/compare/v1.0.0...v1.0.1)
+### Added
+- MIT license
+- Built-in middleware
+- Intergrated phpunit
+- Added routes
+- Added `SNOWFLAKE_HTTP_TOKEN` env variable
+- Added `worker:token` command
+
+
+## [v1.0.1 - 2020-08-24](https://github.com/SepteniTechnology/laravel-snowflake/compare/v1.0.0...v1.0.1)
 ### Added
 - Add facades
 
-## [v1.0.0 - 2020-08-21](https://gitlab.devsep.com/septech-php/septech-laravel/snowflake/-/tree/v1.0.0)
+
+## [v1.0.0 - 2020-08-21](https://github.com/SepteniTechnology/laravel-snowflake/tree/v1.0.0)
 ### Added
 - Release version 1.0 🚀🚀🚀